Some changes to bash, and other takedowns using takedown_check.  takedown_check now uses/returns a float instead of an int for more accurate %s.  Also added a property, skill.bash.ThriKreen_Modifier, which is a multiplier to Thri bash %s.  Touched on springleap to make the takedown_check work right for those skills that use agi.  Now, 50 points of agi or more for the takedown char gives 37.5% towards total, and 50 points in favor of the victim or more gives 0% towards total.  The chance is reduced by 30% to balance this out before adding the agi-based part.

+++ src/prototypes.h
@@ -397,7 +397,7 @@ void maul(P_char, P_char);
 void restrain(P_char, P_char);
 void rescue(P_char, P_char, bool);
 void do_trample(P_char, char *, int);
-int takedown_check(P_char, P_char, int, int, ulong);
+float takedown_check(P_char, P_char, float, int, ulong);
 void do_rearkick(P_char, char *, int);
 void do_whirlwind(P_char, char *, int);
 void kick(P_char, P_char);
